+++ Temperature-Converters/KelvinConverter.cs
@@ -9,30 +9,30 @@ namespace Temperature_Converters
     public class KelvinConverter : ITemperatureConverters
     {
 
-        public object classMain(char fromMode, float Temp)
+        public object classMain(char fromMode, double Temp)
         {
 
             var converterFunctions = new System.Collections.Generic.Dictionary<char, Delegate>();
-            converterFunctions['K'] = new Func<float, float>(fromKelvintoSelf);
-            converterFunctions['F'] = new Func<float, float>(fromFarenheittoSelf);
-            converterFunctions['C'] = new Func<float, float>(fromCelsiustoSelf);
+            converterFunctions['K'] = new Func<double, double>(fromKelvintoSelf);
+            converterFunctions['F'] = new Func<double, double>(fromFarenheittoSelf);
+            converterFunctions['C'] = new Func<double, double>(fromCelsiustoSelf);
 
 
             return converterFunctions[fromMode].DynamicInvoke(Temp);
         }
-        public float fromKelvintoSelf(float Temp)
+        public double fromKelvintoSelf(double Temp)
         {
             
             return Temp;
         }
 
-        public float fromFarenheittoSelf(float Temp)
+        public double fromFarenheittoSelf(double Temp)
         {
             Temp = (Temp + 459.67f)* 5 / 9;
             return Temp;
         }
 
-        public float fromCelsiustoSelf(float Temp)
+        public double fromCelsiustoSelf(double Temp)
         {
 
             return Temp + 273;
